+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - Conversion tool: added parser for Zelcore wallet.
 - Conversion tool: added parser for the BnkToTheFuture platform.
 - Binance parser: added new/re-named operations for Statements. ([#274](https://github.com/BittyTax/BittyTax/issues/274))
+- Binance parser: added config to split BNB quantity across multiple tokens.
 ### Changed
 - Binance parser: performance improvements for large data sets.
 - Coinbase Pro parser: performance improvements for large data sets.